// libs/features/violations/src/lib/shared/models/modify-contributor.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export class ModifyContributor {
  contributorId: number = undefined;
  currentPenaltyAmount: number = undefined;
  excluded: boolean;
  newPenaltyAmount: number = undefined;
  vlContributorId: number = undefined;
}